Move staging logic to separate file
